% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_thumbnail.R
\name{create_ds_thumbnail}
\alias{create_ds_thumbnail}
\title{Create a thumbnail for a data story}
\usage{
create_ds_thumbnail(
  img_path,
  blur_steps = 15,
  max_blur = 4,
  width_blur = 0.4,
  height_blur = 0.34,
  width_crop = 0.58,
  height_crop = 0.62
)
}
\arguments{
\item{img_path}{The path to the screenshot used to generate the thumbnail
(only PNG).}

\item{blur_steps}{Determine the smoothness of the blurring process on the
edges of the thumbnail (large values will requires more computing and then
will take more time. and values lower than 10 are likely to produce a not so
smooth result).}

\item{max_blur}{The intensity of the blur. The blur decreases as it moves
away from the edges.}

\item{width_blur}{The proportion of the image width that will be blurred.
For example, 0.40 means that 40\% of the width (20\% each side) will be
blurred.}

\item{height_blur}{The proportion of the image height that will be blurred.
For example, 0.40 means that 40\% of the height (20\% at the top and 20\% at
the bottom) will be blurred.}

\item{width_crop}{The proportion of the image width that is saved when
cropping, after adding a perspective effect. For example, 0.60 means that
60\% of the image width is saved (20\% of the image is removed on each side).}

\item{height_crop}{The proportion of the image height that is saved when
cropping, after adding a perspective effect. For example, 0.60 means that
60\% of the image height is saved (20\% of the image is at the top and
20\% at the bottom).}
}
\description{
Each data story has a thumbnail on the SNSF data portal data
story main page. This function automatize the creation a thumbnail which
represent a figure selected from the data story with a perspective effect and
some blurring on the edges. This functions is however not magic and will
probably not generate the best image by default for any picture. Each case is
likely to require some fine tuning, but default values are a good starting
point.
}
\keyword{internal}
