% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_blues}
\alias{datastory_blues}
\title{Datastory blues}
\format{
An object of class \code{character} of length 4.
}
\usage{
datastory_blues
}
\description{
'SNSF blue' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_blues)
}
\keyword{datasets}
