% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{datastory_dark_blues}
\alias{datastory_dark_blues}
\title{Datastory dark blues}
\usage{
datastory_dark_blues
}
\description{
'SNSF dark blue' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_dark_blues)
}
