% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{datastory_dark_reds}
\alias{datastory_dark_reds}
\title{Datastory dark reds}
\usage{
datastory_dark_reds
}
\description{
'SNSF dark red' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_dark_reds)
}
