% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_grays}
\alias{datastory_grays}
\title{Datastory grays}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_grays
}
\description{
'SNSF gray' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_grays)
}
\keyword{datasets}
