% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_green}
\alias{datastory_green}
\title{Datastory greens}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_green
}
\description{
'SNSF green' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_green)
}
\keyword{datasets}
