% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{datastory_pal}
\alias{datastory_pal}
\title{Datastory colour palettes generator}
\usage{
datastory_pal(palette = "default", repeat_col = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Character string indicating the datastory colour palette to
use. Only those available in [get_datastory_scheme()] can be used.}

\item{repeat_col}{Logical indicating whether the colour should be repeated
sequentially (`TRUE`) when the palette does not return enough values (the
default). If set to `FALSE`, the number of colours will be extended using
interpolation.}

\item{reverse}{Logical indicating whether the scale should be reversed or
not (the default). Note that when more colours need to be generated, the
the colours are first reversed before before repeated/interpolated (this
allows to get, for a given palette size, consistent colours whatever it is
reversed or not).}
}
\description{
Wrapper function around [get_datastory_scheme()] to transform it
into a palette generator function compatible with
[discrete_scale()] or [scale_fill_gradient()]/[scale_fill_gradient2()].
}
\details{
For more information, see [get_datastory_scheme()] for more
information on the datastory colour palettes available.
}
\keyword{internal}
