% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_reds}
\alias{datastory_reds}
\title{Datastory reds}
\format{
An object of class \code{character} of length 4.
}
\usage{
datastory_reds
}
\description{
'SNSF red' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_reds)
}
\keyword{datasets}
