% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{datastory_scheme_blue_seq}
\alias{datastory_scheme_blue_seq}
\title{(experimental) The datastory SNSF scheme (blue seq)}
\format{
An object of class \code{character} of length 4.
}
\usage{
datastory_scheme_blue_seq
}
\description{
The iterativ interpretation of the YAAY scheme for the data stories.
}
\examples{
 # TODO
}
\keyword{datasets}
