% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{datastory_scheme_green_seq}
\alias{datastory_scheme_green_seq}
\title{SNSF greens (CD 2021)}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_scheme_green_seq
}
\description{
Colors from the new Corporate Design of 2021.
}
\examples{
 # TODO
}
\keyword{datasets}
