% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{datastory_scheme_qualitative}
\alias{datastory_scheme_qualitative}
\title{The default datastory SNSF scheme (qualitative)}
\format{
An object of class \code{character} of length 10.
}
\usage{
datastory_scheme_qualitative
}
\description{
Colors from the new Corporate Design of 2021.
}
\examples{
 # TODO
}
\keyword{datasets}
