% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{datastory_scheme_waffles}
\alias{datastory_scheme_waffles}
\title{Alternative default datastory colour scheme}
\usage{
datastory_scheme_waffles
}
\description{
Same palette as in `datastory_scheme`, but colours at position 2
and 6 are inverted, as well as colours at positions 7 and 12.
}
\examples{
print(datastory_scheme_waffles)
}
