% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_single}
\alias{datastory_single}
\title{Datastory single dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
datastory_single
}
\description{
'SNSF dark blue' colour (100%).
}
\examples{
print(datastory_single)
}
\keyword{datasets}
