% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_violet}
\alias{datastory_violet}
\title{Datastory violets}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_violet
}
\description{
'SNSF violet' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_violet)
}
\keyword{datasets}
