% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_yellow}
\alias{datastory_yellow}
\title{Datastory yellows}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_yellow
}
\description{
'SNSF yellow' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_yellow)
}
\keyword{datasets}
