% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{facet_as_hbar}
\alias{facet_as_hbar}
\title{Turn facets into horizontal bars}
\usage{
facet_as_hbar(facet, ncol = 1L, ...)
}
\arguments{
\item{facet}{Variable to use for faceting (must be the same as the one set
for y-axis in the aesthetics of the plot).}

\item{ncol}{Integer indicating the number of facet columns (default is 1).}

\item{...}{Further arguments passed to [ggplot2::facet_wrap()].}
}
\value{
An object of class `<ggproto>`.
}
\description{
Turn facets into horizontal bars
}
\examples{
library(ggplot2)

ggplot(mpg, aes(y = class, fill = class)) +
  geom_bar() +
  scale_fill_datastory(guide = NULL) +
  # You need to modify the x-axis with the function below for nice alignment
  scale_x_facet_as_hbar() +
  facet_as_hbar(~class) +
  get_datastory_theme(
    # Make sure to indicate the theme you are using facet as bars
    facet_as_hbar = TRUE
  )
}
