% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{format_when_plotly}
\alias{format_when_plotly}
\title{Make text bold for plotly with HTML}
\usage{
format_when_plotly(x, is_plotly = TRUE, make_bold = TRUE, make_italic = FALSE)
}
\arguments{
\item{x}{Character string to be formatted}

\item{is_plotly}{Whether to apply the formatting or not}

\item{make_bold}{Bold or not?}

\item{make_italic}{Italic or not?}
}
\description{
\code{plotly::ggplotly()} is convenient but does not support every aspect of
a ggplot plot when converting it into a plotly plot. One of these aspects is
the \code{fontface} parameter of \code{geom_text}. The font face can still be
controlled for plotly output when wrapping HTML tags around the actual text
in \code{geom_text}, this function does this when the wished output string
should be formatted for plotly.
}
\examples{
 format_when_plotly("hallo!", is_plotly = TRUE, make_bold = TRUE)
}
