% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_datastory_theme}
\alias{get_datastory_theme}
\title{(experimental) Get the ggplot2 theme for data story plots}
\usage{
get_datastory_theme(
  legend_position = "top",
  legend_margin = margin(0, 0, 0, 0),
  gridline_axis = c(),
  title_axis = c(),
  text_axis = c("x", "y"),
  tick_axis = c(),
  remove_plot_margin = FALSE,
  legend_key_size = c()
)
}
\arguments{
\item{legend_position}{TODO}

\item{legend_margin}{TODO}

\item{gridline_axis}{TODO}

\item{title_axis}{TODO}

\item{text_axis}{TODO}

\item{tick_axis}{TODO}

\item{remove_plot_margin}{TODO}

\item{legend_key_size}{TODO}
}
\description{
The data plot theme is specially tailored for the clean look of the data
stories. By default, no axis titles and no grid lines are included, they can
be added when needed.
}
\examples{
 # TODO
}
