% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{print_num}
\alias{print_num}
\title{Print numbers as characters with language specific format}
\usage{
print_num(x, lang = "en")
}
\arguments{
\item{x}{The numeric to be transformed}

\item{lang}{Output format language, available is "en", "de", or "fr"}
}
\description{
Function to format a given number with a specified language-specific
formatting (thousand separator and decimal separator)
}
\examples{
 translate_research_area("SSH", target_lang = "de", abbr_or_long = "long")
}
