% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_area_formatting.R
\name{reformat_research_area}
\alias{reformat_research_area}
\title{(experimental) Translating research area names}
\usage{
reformat_research_area(
  research_area,
  from_lang,
  to_lang,
  from_format,
  to_format,
  output_as_factor = FALSE
)
}
\arguments{
\item{research_area}{Research area to be translated as English abbreviation,
either "SSH", "MINT", or "LS".}

\item{from_lang}{The language of the input ("de", "en", or "fr").}

\item{to_lang}{The language of the output ("de", "en", or "fr").}

\item{from_format}{The format of the input ("short" or "long").}

\item{to_format}{The format of the output ("short" or "long").}

\item{output_as_factor}{Whether you want the output as factor (\code{TRUE}) or as
character (\code{FALSE}, default). If this is set to \code{TRUE}, levels will be ordered as
follows: SSH, MINT, and LS.}
}
\description{
Function flexibly reformat research area names from/to long or short formats,
as well as from/to any of the three languages used (English, German, and
French). This function can be used flexibly to translate a research area,
reformat within the same language, or do both at the same time.
}
\examples{
 reformat_research_area(
     "SSH",
     from_lang = "en",
     to_lang = "de",
     from_format = "short",
     to_format = "long"
 )
}
