% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_facet_as_hbar}
\alias{scale_x_facet_as_hbar}
\title{Modify x-axis scale to render nicely facets as bars}
\usage{
scale_x_facet_as_hbar()
}
\value{
An object of class `<ggproto>`.
}
\description{
Modify x-axis scale to render nicely facets as bars
}
\examples{
library(ggplot2)

ggplot(mpg, aes(y = class, fill = class)) +
  geom_bar() +
  scale_fill_datastory(guide = NULL) +
  # You need to modify the x-axis with the function below for nice alignment
  scale_x_facet_as_hbar() +
  facet_as_hbar(~class) +
  get_datastory_theme(
    # Make sure to indicate the theme you are using facet as bars
    facet_as_hbar = TRUE
  )
}
