% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_dark_reds}
\alias{snsf_dark_reds}
\title{"SNF Dunkelrot" colors (100\%, 70\%, 30\%, 10\%)
CD 2021+
Primary colors.}
\format{
An object of class \code{character} of length 4.
}
\usage{
snsf_dark_reds
}
\description{
"SNF Dunkelrot" colors (100\%, 70\%, 30\%, 10\%)
CD 2021+
Primary colors.
}
\examples{
 print(snsf_dark_reds)
}
\keyword{datasets}
