% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_grays}
\alias{snsf_grays}
\title{"SNF Grau" colors (100\%, 50\%)
CD 2021+
Secondary colors.}
\format{
An object of class \code{character} of length 2.
}
\usage{
snsf_grays
}
\description{
"SNF Grau" colors (100\%, 50\%)
CD 2021+
Secondary colors.
}
\examples{
 print(snsf_grays)
}
\keyword{datasets}
