% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_reds}
\alias{snsf_reds}
\title{"SNF Rot" colors (100\%, 70\%, 30\%, 10\%)}
\format{
An object of class \code{character} of length 4.
}
\usage{
snsf_reds
}
\description{
Primary colors CD 2021+
}
\examples{
 print(snsf_reds)
}
\keyword{datasets}
