% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_violet}
\alias{snsf_violet}
\title{SNF Violett" colors (100\%, 50\%)}
\format{
An object of class \code{character} of length 2.
}
\usage{
snsf_violet
}
\description{
Secondary colors CD 2021+
}
\examples{
 print(snsf_violet)
}
\keyword{datasets}
