% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{translate_research_area}
\alias{translate_research_area}
\title{(experimental) Translating research area names}
\usage{
translate_research_area(
  research_area = "",
  target_lang = "en",
  abbr_or_long = "abbr"
)
}
\arguments{
\item{research_area}{Research area to be translated as English abbreviation,
either "SSH", "MINT", or "LS".}

\item{target_lang}{Target language as abbreviation ("de", "en", or "fr").}

\item{abbr_or_long}{Whether the output string should be an abbreviation or
the whole, long name (possible values: "abbr" or "long").}
}
\description{
Function to translate the English research areas (SSH, MINT, LS) names into
the corresponding language versions (de, fr, en) as abbreviations or long
format
}
\examples{
 translate_research_area("SSH", target_lang = "de", abbr_or_long = "long")
}
