% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_dark_blues}
\alias{datastory_dark_blues}
\title{Datastory dark blues}
\format{
An object of class \code{character} of length 4.
}
\usage{
datastory_dark_blues
}
\description{
'SNSF dark blue' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_dark_blues)
}
\keyword{datasets}
