% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_dark_reds}
\alias{datastory_dark_reds}
\title{Datastory dark reds}
\format{
An object of class \code{character} of length 4.
}
\usage{
datastory_dark_reds
}
\description{
'SNSF dark red' colours (100%, 70%, 30%, 10%).
}
\examples{
print(datastory_dark_reds)
}
\keyword{datasets}
