% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{datastory_scheme}
\alias{datastory_scheme}
\title{Default Datastory colour schemes}
\format{
An object of class \code{character} of length 12.
}
\usage{
datastory_scheme
}
\description{
SNSF main palette composed of 12 colours. The first 6 colours
are based on SNSF secondary colours (position 1, 3, and 4), dark blue 100%
(position  2), red 100% (position 5), and blue 100% (position 6). The 6
remaining colours are a lighter version of the first 6.
}
\examples{
print(datastory_scheme)
}
\keyword{datasets}
