% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{datastory_scheme_yellow_seq}
\alias{datastory_scheme_yellow_seq}
\title{SNSF yellows (CD 2021)}
\format{
An object of class \code{character} of length 2.
}
\usage{
datastory_scheme_yellow_seq
}
\description{
SNSF yellows (CD 2021)
}
\examples{
 # TODO
}
\keyword{datasets}
