% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/scales.R
\name{get_datastory_scheme}
\alias{get_datastory_scheme}
\title{Get your datastory SNSF scheme}
\usage{
get_datastory_scheme(
  n_col = NULL,
  palette = "default",
  repeat_col = TRUE,
  reverse = FALSE
)

get_datastory_scheme(
  n_col = NULL,
  palette = "default",
  repeat_col = TRUE,
  reverse = FALSE
)
}
\arguments{
\item{n_col}{The number of colours to return. If it exceeds the number of
colours in the chosen palette, additional colours are interpolated. Do not
set this argument if you want to get the original number of colours in the
chosen palette.}

\item{palette}{Choose between "default" and "waffles".}

\item{repeat_col}{Whether to repeat the colour scheme when there are more
colours requested (with `n_col`) than the colour scheme has to offer. When
set to `FALSE`, additional colours are interpolated.}

\item{reverse}{Whether the palette should be returned in reversed order. If
`n_col` is set and is smaller than the total number of colours in the
palette, the function takes the first `n_col` colours in the palette and
reverses their order. It does not reverse the whole palette and then take the
first `n_col` colours.}
}
\description{
palettes : default (qualitative), qualitative, blue_seq,
green_seq, yellow_seq, gray_seq.

Choose between different datastory colour palettes and get
additional colours through interpolation when needed.
}
\examples{
 # TODO
# Default colours for charts
get_datastory_scheme()
# Colours for waffle charts
get_datastory_scheme(palette = "waffles")
}
