% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_datastory_scheme}
\alias{get_datastory_scheme}
\title{(experimental) Get your datastory SNSF scheme}
\usage{
get_datastory_scheme(palette = "default", n_col = NULL)
}
\description{
palettes : default (qualitative), qualitative, blue_seq, green_seq,
yellow_seq, gray_seq.
}
\examples{
 # TODO
}
