% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_datastory_theme}
\alias{get_datastory_theme}
\title{Get the ggplot2 theme for data story plots}
\usage{
get_datastory_theme(
  legend_position = "top",
  legend_margin = ggplot2::margin(0, 0, 0, 0),
  gridline_axis = c(),
  title_axis = c(),
  text_axis = c("x", "y"),
  tick_axis = c(),
  remove_plot_margin = FALSE,
  legend_key_size = c(),
  family = "Theinhardt"
)
}
\arguments{
\item{legend_position}{Choose the position of the legend in regards to your
plot.}

\item{legend_margin}{ggplot \code{legend.margin} theme element.}

\item{gridline_axis}{TODO}

\item{title_axis}{Which axis titles should be shown, default X and Y:
\code{c("x", "y")}}

\item{text_axis}{Which axis text should be shown, default X and Y:
\code{c("x", "y")}}

\item{tick_axis}{On which axis the axis ticks should be displayed.}

\item{remove_plot_margin}{Whether all margin around the plot should be
removed. Default \code{FALSE}.}

\item{legend_key_size}{ggplot legend.key.width and legend.key.height theme
item: vector with first element width (mm) and second element height (mm).}

\item{family}{the name of the font to use to display text. Theinhardt is
used as default.}
}
\description{
The data plot theme is specially tailored for the clean look of
the data stories. By default, no axis titles and no grid lines are included,
they can be added when needed.
}
\examples{
 # TODO
}
