% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_ggiraph_tooltip_css}
\alias{get_ggiraph_tooltip_css}
\title{Get ggiraph tooltip CSS}
\usage{
get_ggiraph_tooltip_css(family = "Source Sans Pro")
}
\arguments{
\item{family}{the name of the font to use to display text. Source Sans Pro is
used as default.}
}
\description{
Get uniform tooltip CSS style for ggiraph graphs.
}
