% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_ggiraph_tooltip_css}
\alias{get_ggiraph_tooltip_css}
\title{Get ggiraph tooltip CSS}
\usage{
get_ggiraph_tooltip_css()
}
\description{
Get uniform tooltip CSS style for ggiraph graphs.
}
\examples{
 ggiraph::girafe(
 ggobj = ggplot(mtcars),
 options = list(
   ggiraph::opts_tooltip(
     css = get_ggiraph_tooltip_css()
   )
 ))
}
