% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_number.R
\name{print_num}
\alias{print_num}
\title{Print numbers as characters with language specific format}
\usage{
print_num(x, lang = "en")
}
\arguments{
\item{x}{The numeric to be transformed}

\item{lang}{Output format language, available is "en", "de", or "fr"}
}
\description{
Function to format a given number with a specified language-specific
formatting (thousand separator and decimal separator)
}
\examples{
 print_num(10000, "en")
 print_num(10000, "fr")
 print_num(10000, "de")
 print_num(pi, "en")
 print_num(pi, "fr")
 print_num(pi, "de")
 print_num(10000 * pi, "en")
 print_num(10000 * pi, "fr")
 print_num(10000 * pi, "de")
}
