% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_black}
\alias{snsf_black}
\title{"SNF Schwarzblau" primary color (100\%)
CD 2021+
Primary colors.}
\format{
An object of class \code{character} of length 1.
}
\usage{
snsf_black
}
\description{
"SNF Schwarzblau" primary color (100\%)
CD 2021+
Primary colors.
}
\examples{
 print(snsf_black)
}
\keyword{datasets}
