% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_blues}
\alias{snsf_blues}
\title{"SNF Blau" colors (100\%, 70\%, 30\%, 10\%)
CD 2021+
Primary colors.}
\format{
An object of class \code{character} of length 4.
}
\usage{
snsf_blues
}
\description{
"SNF Blau" colors (100\%, 70\%, 30\%, 10\%)
CD 2021+
Primary colors.
}
\examples{
 print(snsf_blues)
}
\keyword{datasets}
