% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{snsf_green}
\alias{snsf_green}
\title{"SNF Grün" colors (100\%, 50\%)}
\format{
An object of class \code{character} of length 2.
}
\usage{
snsf_green
}
\description{
Secondary colors CD 2021+
}
\examples{
 print(snsf_green)
}
\keyword{datasets}
